add all_integer and all_integer_predictors

+++ R/misc.R
@@ -9,7 +9,7 @@ get_types <- function(x) {
       character = "nominal",
       ordered = "ordered",
       factor = "nominal",
-      integer = "numeric",
+      integer = "integer",
       numeric = "numeric",
       double = "numeric",
       Surv = "censored",
